Зайцев Артем. Задача 1. Вариант 29. Технология TBB. Маркировка компонент на бинарном изображении. (#372)
## Principle
The Two Pass algorithm with 8-connectedness neighbourhood mask is used.
The program sequentially iterates through the pixels and decides whether
the next one is the background, part of an existing connectivity
component, or the first one in a new component, labeling each pixel with
the name of the component to which it belongs.
### Detailed
I) Iterate through the pixels:
1. Check pixel's color. If it's black, label it as background (label
`0`) and go to the step 3. Otherwise go to step 2.
2. Get the labels of the neighbours of the current pixel.
a) If all ones are zero-labeled, mark current pixel with the label
unused before and go to the step 3.
b) Otherwise, set the minimal label for current pixel, and write down
all neighbouring pixels ad equivalent for each other. For those
operations Union-Find (disjoint set) data structure is used.
3. If there are unprocessed pixels, go to the next one and do the step
1. Otherwise go to the step II).

II) Calculate true labels for each pixel. Assign true labels instead of
temporary ones.
### Multiprocessing scheme
Input image is being splitted into `num_threads` chunks. For each chunk
the algorithm performs both the the first and the second scan. After the
end of the parallel region chunks are being united into a single label
map with the shifted labels. At the edges of previous chunks new
equivalencies map is being built. After that program performs
replacements and renamings to make labels sequence continuous.
## Input & Output
**Input:** width and height of the image and its flat representation as
`std::vector`.
**Output:** flat `std::vector` of labels. Its two-dimensional
representation has the same size as the input image do.